+++ src/servers/ManageServersRow.tsx
@@ -27,7 +27,11 @@ const ManageServersRow: FCWithDeps<ManageServersRowProps, ManageServersRowDeps> 
         <Table.Cell columnName="Auto-connect">
           {server.autoConnect && (
             <>
-              <FontAwesomeIcon icon={checkIcon} className="tw:text-brand" id="autoConnectIcon" />
+              <FontAwesomeIcon
+                icon={checkIcon}
+                className="tw:text-lm-brand tw:dark:text-dm-brand"
+                id="autoConnectIcon"
+              />
               <UncontrolledTooltip target="autoConnectIcon" placement="right">
                 Auto-connect to this server
               </UncontrolledTooltip>